feat(sxhkd): add initial sxhkdrc
Add a minimal resource configuration file for the simple X11 hotkey
daemon, the best hotkey daemon with a name everyone struggles to
pronounce!

Plenty in here in still reliant on scripts that are deserving of their
own repository, which might soon be created and made public. It's very
setup dependent, however.