chore(zsh): remove hardcoded libva driver name
For some reason, in the past I needed to hardcode the libva driver name
to i965 so my Intel integrated graphics chip could be used for hardware
acceleration through VA-API.

Now that I'm primarily using different machines and this can be set per
machine rather easily, this is being removed. It's getting in the way of
using other UHD for hardware acceleration, after all.