feat(xmonad): add modifier to hide windows automatically
Add a modifier to hide the Nth window once the workspace reaches a
specific number of windows. This is a bit of hack, as ideally this would
behave just like "limit" and be able to take a single parameter, with
the window to be hidden being the one focused immediately before a new
window was added to the stack.

This is pending some refactoring, but for now it works.