feat(nvim): add plugins & respective configuration
Add plugins to be loaded by packer.nvim, as well as their respective
configuration. This should somewhat mirror the old configuration as
existed in this repository written in VimL, which used vim-plug for
plugin management.

This makes init.lua a hefty file, with config that might be removed in
the coming weeks.