fix(nvim): properly create autoformat autocmd
Create the autocommand that enables paragraph autoformatting the way it
was meant to be in the first place. Maybe it makes more sense for people
who understand VimL, but 'pattern' here is (also?) used for FileType's
argument.

Seems the function I was using as callback didn't quite work, either, so
it was changed to a regular Vim command.