chore(nvim): clear default style for NormalFloat
Make it so NormalFloat, the highlight group for floating windows, which
is now used by FZF commands from fzf.vim, doesn't have any styling.

A side effect of this is that popup windows, which are also using this
highlight group, will no longer have clearly distinguishable borders
separating it from regular text. Not too much of a bother currently.