feat(tmux): set window title to include session name
Change the terminal emulator window title to include the tmux session
name in addition to a reference to tmux. This makes it easier to
identify terminal windows by the tmux session that it is attached to.

The XMonad configuration was changed to use the application name to
identify the terminal named scratchpad, as the window title is no longer
unique.